requirement.txt
Package                      Version
---------------------------- --------
aiohttp                      3.9.5
aiosignal                    1.3.1
annotated-types              0.7.0
anyio                        4.4.0
attrs                        23.2.0
cachetools                   5.3.3
certifi                      2024.2.2
charset-normalizer           3.3.2
colorama                     0.4.6
dataclasses-json             0.6.6
frozenlist                   1.4.1
google-ai-generativelanguage 0.6.4
google-api-core              2.19.0
google-api-python-client     2.131.0
google-auth                  2.29.0
google-auth-httplib2         0.2.0
google-generativeai          0.5.4
googleapis-common-protos     1.63.0
greenlet                     3.0.3
grpcio                       1.64.0
grpcio-status                1.62.2
h11                          0.14.0
httpcore                     1.0.5
httplib2                     0.22.0
httpx                        0.27.0
idna                         3.7
jsonpatch                    1.33
jsonpointer                  2.4
langchain                    0.2.1
langchain-community          0.2.1
langchain-core               0.2.1
langchain-google-genai       1.0.5
langchain-text-splitters     0.2.0
langsmith                    0.1.63
marshmallow                  3.21.2
multidict                    6.0.5
mypy-extensions              1.0.0
numpy                        1.26.4
ollama                       0.2.0
orjson                       3.10.3
packaging                    23.2
pillow                       10.3.0
pip                          24.0
proto-plus                   1.23.0
protobuf                     4.25.3
pyasn1                       0.6.0
pyasn1_modules               0.4.0
pydantic                     2.7.2
pydantic_core                2.18.3
pyparsing                    3.1.2
python-dotenv                1.0.1
PyYAML                       6.0.1
requests                     2.32.2
rsa                          4.9
sniffio                      1.3.1
SQLAlchemy                   2.0.30
tenacity                     8.3.0
tqdm                         4.66.4
typing_extensions            4.12.0
typing-inspect               0.9.0
uritemplate                  4.1.1
urllib3                      2.2.1
yarl                         1.9.4
